refactor: GLM adapter 분리 + DRY 헬퍼 추출 + GAP 수정
검증팀 GAP 탐지 결과 5건 수정:

1. glm_adapter.py 분리 — openai_adapter.py에서 GLM 코드 독립
2. _helpers.py 추출 — _mask_key, _upsert_env, is_glm_key 공유
3. GLM 키 탐지 패턴 통일 — is_glm_key() 단일 소스
4. /auth add에 ZhipuAI 프로바이더 추가
5. SubAgent provider 상속 — settings.model + _resolve_provider() 전달

2814 passed, 0 failed
